Uploaded Code for Initial Release
This is the first release, so expect a few bugs. I have done my best to unit test the backend and provide a user interface to alert when there is possible data loss so that it may be avoided. I would love to hear any bug reports, concerns, and suggestions you may have.